feat: Piece 객체에서 팩토리 메소드를 PieceFactory 객체의 메소드로 변경
