+++ src/main/java/kr/codesqaud/chessgame/ChessApplication.java
@@ -2,12 +2,10 @@ package kr.codesqaud.chessgame;
 
 import kr.codesqaud.chessgame.chess.ChessGame;
 
-import java.io.FileInputStream;
-
 public class ChessApplication {
 
     public static void main(String[] args) {
-        final ChessGame chessGame = ChessGame.getInstance();
-        chessGame.start(System.in);
+        final ChessGame chessGame = new ChessGame();
+        chessGame.run(System.in);
     }
 }
