+++ src/main/java/kr/codesqaud/chessgame/pieces/Color.java
@@ -1,0 +1,6 @@
+package kr.codesqaud.chessgame.pieces;
+
+public enum Color {
+    WHITE, BLACK, NOCOLOR;
+}
+
