+++ src/main/java/kr/codesqaud/chessgame/pieces/Position.java
@@ -61,7 +61,6 @@ public class Position {
         return rank - 1;
     }
 
-
     @Override
     public boolean equals(final Object o) {
         if (this == o) {
